requirements-dev.txt


ase
cython
h5py
matplotlib
mendeleev
numpy
pandas
paramiko
pexpect
pymatgen
python-nomad
pyyaml
qmpy-rester
scp
seaborn
setuptools
sqlalchemy
tensorflow
uncertainties
nomad-lab
loguru
matminer


autopep8
coverage
doc8
flake8
nose
pre-commit
pyaml
pyblack
pytest
sphinx
tox


